+++ src/Products/PloneMeeting/events.py
@@ -20,6 +20,7 @@ from imio.helpers.content import richtextval
 from imio.helpers.security import fplog
 from imio.helpers.xhtml import storeImagesLocally
 from OFS.ObjectManager import BeforeDeleteException
+from OFS.interfaces import IObjectWillBeAddedEvent
 from persistent.list import PersistentList
 from persistent.mapping import PersistentMapping
 from plone import api
@@ -1448,8 +1449,8 @@ def onFacetedGlobalSettingsChanged(folder, event):
     _notifyContainerModified(folder)
 
 
-def onCategoryWillBeRemoved(category, event):
-    '''Checks if the current p_category can be deleted:
+def onCategoryWillBeMoved(category, event):
+    '''Checks if the current p_category can be moved (renamed) or deleted:
       - it can not be linked to an existing meetingItem (normal item,
         recurring item or item template);
       - it can not be used in field 'category_mapping_when_cloning_to_other_mc'
@@ -1457,6 +1458,10 @@ def onCategoryWillBeRemoved(category, event):
     # If we are trying to remove the whole Plone Site, bypass this hook.
     # bypass also if we are in the creation process
     if event.object.meta_type == 'Plone Site':
+        return
+
+    # also called when new category created
+    if IObjectWillBeAddedEvent.providedBy(event):
         return
 
     tool = api.portal.get_tool('portal_plonemeeting')