+++ CHANGES.rst
@@ -143,6 +143,8 @@ Changelog
   that will check that there are no more items `Returned to proposing group` when
   publishing decisions.
   [gbastien]
+- Make sure a `MeetingCategory` can not be renamed if it is used.
+  [gbastien]
 
 4.2rc34 (2022-09-29)
 --------------------